made modifications with help from a friend and officially made a bmi calculator

+++ task_2.cpp
@@ -9,8 +9,8 @@
 int main(){
 
     std::string name;
-    int weight;
-    int height;
+    double weight;
+    double height;
     
     
     std::cout << " what is your full name?" << std::endl;
